feat: change to Double
- Represent levels as floor of userReputationScore.

+++ app/src/main/java/com/github/se/cyrcle/model/user/User.kt
@@ -16,7 +16,7 @@ data class UserPublic(
     val profilePictureCloudPath: String? = "",
     // val accountCreationDate: Timestamp? = null,
     // val numberOfContributedSpots: Int = 0,
-    val userReputationScore: Int = 0
+    val userReputationScore: Double = 0.0
 )
 
 /**
